Simple demonstration of variables built in.
There is no need for the code to read variables in text as they can be used through the js function `${var}` already.

This is a simple demonstration to show that variables in text behave the same as plain text, when used in the same way as the "Lamia" example used here.